Children interface is now ReactNode, theme utility methods are deprecated in favour of plugin methods
This fixes a type incompatibility where using `guard()` and returning
a React or DOM element would not pass type checking. The Children
interface now extends ReactNode for backwards compatibility and is
marked as depreacated. Use ReactNode directly in future.

Duplicated markdown utility functions in the theme are now deprecated
in favour of their equivalents in the plugin package.